# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carbon Footprint Awareness Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ActivityCategory = typeof ActivityCategory[keyof typeof ActivityCategory];


export const ActivityCategory = {
  transport: 'transport',
  energy: 'energy',
  food: 'food',
  shopping: 'shopping',
} as const;

export interface Activity {
  id: number;
  category: ActivityCategory;
  activityType: string;
  amount: number;
  unit: string;
  /** CO2 equivalent in kilograms */
  co2Kg: number;
  /** @nullable */
  notes?: string | null;
  date: string;
  createdAt: string;
}

export type ActivityInputCategory = typeof ActivityInputCategory[keyof typeof ActivityInputCategory];


export const ActivityInputCategory = {
  transport: 'transport',
  energy: 'energy',
  food: 'food',
  shopping: 'shopping',
} as const;

export interface ActivityInput {
  category: ActivityInputCategory;
  /** @minLength 1 */
  activityType: string;
  /** @minimum 0 */
  amount: number;
  /** @minLength 1 */
  unit: string;
  /** @minimum 0 */
  co2Kg: number;
  notes?: string;
  date: string;
}

export interface ActivityUpdate {
  /** @minimum 0 */
  amount?: number;
  /** @minimum 0 */
  co2Kg?: number;
  notes?: string;
  date?: string;
}

export type ActionCategory = typeof ActionCategory[keyof typeof ActionCategory];


export const ActionCategory = {
  transport: 'transport',
  energy: 'energy',
  food: 'food',
  shopping: 'shopping',
  lifestyle: 'lifestyle',
} as const;

export type ActionDifficulty = typeof ActionDifficulty[keyof typeof ActionDifficulty];


export const ActionDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface Action {
  id: number;
  title: string;
  description: string;
  category: ActionCategory;
  co2SavedKgPerYear: number;
  difficulty: ActionDifficulty;
  completed: boolean;
  /** @nullable */
  completedAt?: string | null;
}

export interface ActionToggle {
  completed: boolean;
}

export interface Dashboard {
  totalCo2ThisMonth: number;
  totalCo2ThisWeek: number;
  totalCo2AllTime: number;
  /** Global average monthly CO2 in kg (normalized to 333 kg/month) */
  averagePersonCo2Monthly: number;
  /** User's footprint as percentage of global average */
  percentageVsAverage: number;
  activitiesThisMonth: number;
  actionsCompleted: number;
  co2SavedByActions: number;
  /** Consecutive days with at least one activity logged */
  streak: number;
}

export interface CategoryBreakdown {
  category: string;
  co2Kg: number;
  percentage: number;
  count: number;
}

export interface TrendPoint {
  week: string;
  label: string;
  co2Kg: number;
}

export interface Recommendation {
  category: string;
  title: string;
  description: string;
  potentialSavingKg: number;
  priority: number;
}

export interface Profile {
  id: number;
  name: string;
  /** @nullable */
  location: string | null;
  monthlyGoalKg: number;
  onboardingCompleted: boolean;
  createdAt: string;
}

export interface ProfileUpdate {
  /** @minLength 1 */
  name?: string;
  location?: string;
  /** @minimum 1 */
  monthlyGoalKg?: number;
  onboardingCompleted?: boolean;
}

export type ListActivitiesParams = {
category?: ListActivitiesCategory;
limit?: number;
};

export type ListActivitiesCategory = typeof ListActivitiesCategory[keyof typeof ListActivitiesCategory];


export const ListActivitiesCategory = {
  transport: 'transport',
  energy: 'energy',
  food: 'food',
  shopping: 'shopping',
} as const;

export type ListActionsParams = {
category?: ListActionsCategory;
completed?: boolean;
};

export type ListActionsCategory = typeof ListActionsCategory[keyof typeof ListActionsCategory];


export const ListActionsCategory = {
  transport: 'transport',
  energy: 'energy',
  food: 'food',
  shopping: 'shopping',
  lifestyle: 'lifestyle',
} as const;

export type GetBreakdownParams = {
period?: GetBreakdownPeriod;
};

export type GetBreakdownPeriod = typeof GetBreakdownPeriod[keyof typeof GetBreakdownPeriod];


export const GetBreakdownPeriod = {
  week: 'week',
  month: 'month',
  year: 'year',
} as const;

export type GetTrendsParams = {
weeks?: number;
};