# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Carbon Footprint Awareness Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: activities
    description: Carbon activity tracking
  - name: actions
    description: Carbon reduction actions
  - name: insights
    description: Personalized insights and dashboard
  - name: profile
    description: User profile and settings

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List all activity logs
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [transport, energy, food, shopping]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Log a new carbon activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
        "400":
          description: Invalid input

  /activities/{id}:
    get:
      operationId: getActivity
      tags: [activities]
      summary: Get a specific activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
        "404":
          description: Not found
    patch:
      operationId: updateActivity
      tags: [activities]
      summary: Update an activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityUpdate"
      responses:
        "200":
          description: Updated activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
        "404":
          description: Not found
    delete:
      operationId: deleteActivity
      tags: [activities]
      summary: Delete an activity log
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /actions:
    get:
      operationId: listActions
      tags: [actions]
      summary: List reduction actions
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [transport, energy, food, shopping, lifestyle]
        - name: completed
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of actions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Action"

  /actions/{id}/complete:
    post:
      operationId: completeAction
      tags: [actions]
      summary: Mark an action as completed or uncompleted
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActionToggle"
      responses:
        "200":
          description: Updated action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Action"
        "404":
          description: Not found

  /insights/dashboard:
    get:
      operationId: getDashboard
      tags: [insights]
      summary: Get dashboard summary with totals and trends
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"

  /insights/breakdown:
    get:
      operationId: getBreakdown
      tags: [insights]
      summary: Get emissions breakdown by category
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, year]
            default: month
      responses:
        "200":
          description: Breakdown by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryBreakdown"

  /insights/trends:
    get:
      operationId: getTrends
      tags: [insights]
      summary: Get weekly emission trends over time
      parameters:
        - name: weeks
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Weekly trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendPoint"

  /insights/recommendations:
    get:
      operationId: getRecommendations
      tags: [insights]
      summary: Get top personalized recommendations based on user data
      responses:
        "200":
          description: Top 3 personalized recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recommendation"

  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Activity:
      type: object
      required: [id, category, activityType, amount, unit, co2Kg, date, createdAt]
      properties:
        id:
          type: integer
        category:
          type: string
          enum: [transport, energy, food, shopping]
        activityType:
          type: string
        amount:
          type: number
        unit:
          type: string
        co2Kg:
          type: number
          description: CO2 equivalent in kilograms
        notes:
          type: ["string", "null"]
        date:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time

    ActivityInput:
      type: object
      required: [category, activityType, amount, unit, co2Kg, date]
      properties:
        category:
          type: string
          enum: [transport, energy, food, shopping]
        activityType:
          type: string
          minLength: 1
        amount:
          type: number
          minimum: 0
        unit:
          type: string
          minLength: 1
        co2Kg:
          type: number
          minimum: 0
        notes:
          type: string
        date:
          type: string
          format: date

    ActivityUpdate:
      type: object
      properties:
        amount:
          type: number
          minimum: 0
        co2Kg:
          type: number
          minimum: 0
        notes:
          type: string
        date:
          type: string
          format: date

    Action:
      type: object
      required: [id, title, description, category, co2SavedKgPerYear, difficulty, completed]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [transport, energy, food, shopping, lifestyle]
        co2SavedKgPerYear:
          type: number
        difficulty:
          type: string
          enum: [easy, medium, hard]
        completed:
          type: boolean
        completedAt:
          type: ["string", "null"]

    ActionToggle:
      type: object
      required: [completed]
      properties:
        completed:
          type: boolean

    Dashboard:
      type: object
      required: [totalCo2ThisMonth, totalCo2ThisWeek, totalCo2AllTime, averagePersonCo2Monthly, percentageVsAverage, activitiesThisMonth, actionsCompleted, co2SavedByActions, streak]
      properties:
        totalCo2ThisMonth:
          type: number
        totalCo2ThisWeek:
          type: number
        totalCo2AllTime:
          type: number
        averagePersonCo2Monthly:
          type: number
          description: Global average monthly CO2 in kg (normalized to 333 kg/month)
        percentageVsAverage:
          type: number
          description: User's footprint as percentage of global average
        activitiesThisMonth:
          type: integer
        actionsCompleted:
          type: integer
        co2SavedByActions:
          type: number
        streak:
          type: integer
          description: Consecutive days with at least one activity logged

    CategoryBreakdown:
      type: object
      required: [category, co2Kg, percentage, count]
      properties:
        category:
          type: string
        co2Kg:
          type: number
        percentage:
          type: number
        count:
          type: integer

    TrendPoint:
      type: object
      required: [week, co2Kg, label]
      properties:
        week:
          type: string
          format: date
        label:
          type: string
        co2Kg:
          type: number

    Recommendation:
      type: object
      required: [category, title, description, potentialSavingKg, priority]
      properties:
        category:
          type: string
        title:
          type: string
        description:
          type: string
        potentialSavingKg:
          type: number
        priority:
          type: integer

    Profile:
      type: object
      required: [id, name, location, monthlyGoalKg, onboardingCompleted, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: ["string", "null"]
        monthlyGoalKg:
          type: number
        onboardingCompleted:
          type: boolean
        createdAt:
          type: string
          format: date-time

    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        location:
          type: string
        monthlyGoalKg:
          type: number
          minimum: 1
        onboardingCompleted:
          type: boolean