# lib/api-zod/src/generated/types/actionToggle.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carbon Footprint Awareness Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ActionToggle {
  completed: boolean;
}