# lib/api-zod/src/generated/types/activityInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carbon Footprint Awareness Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityInputCategory } from './activityInputCategory';

export interface ActivityInput {
  category: ActivityInputCategory;
  /** @minLength 1 */
  activityType: string;
  /** @minimum 0 */
  amount: number;
  /** @minLength 1 */
  unit: string;
  /** @minimum 0 */
  co2Kg: number;
  notes?: string;
  date: Date;
}