# lib/api-zod/src/generated/types/activityUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carbon Footprint Awareness Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityUpdate {
  /** @minimum 0 */
  amount?: number;
  /** @minimum 0 */
  co2Kg?: number;
  notes?: string;
  date?: Date;
}