# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carbon Footprint Awareness Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Dashboard {
  totalCo2ThisMonth: number;
  totalCo2ThisWeek: number;
  totalCo2AllTime: number;
  /** Global average monthly CO2 in kg (normalized to 333 kg/month) */
  averagePersonCo2Monthly: number;
  /** User's footprint as percentage of global average */
  percentageVsAverage: number;
  activitiesThisMonth: number;
  actionsCompleted: number;
  co2SavedByActions: number;
  /** Consecutive days with at least one activity logged */
  streak: number;
}