# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carbon Footprint Awareness Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: number;
  name: string;
  /** @nullable */
  location: string | null;
  monthlyGoalKg: number;
  onboardingCompleted: boolean;
  createdAt: Date;
}