# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Carbon Footprint Awareness Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileUpdate {
  /** @minLength 1 */
  name?: string;
  location?: string;
  /** @minimum 1 */
  monthlyGoalKg?: number;
  onboardingCompleted?: boolean;
}